+++ src/main/java/com/kkyeer/debugger/to/uml/ui/StackFrameControlPanel.java
@@ -2,7 +2,10 @@ package com.kkyeer.debugger.to.uml.ui;
 
 import com.intellij.debugger.engine.JavaStackFrame;
 import com.intellij.debugger.ui.impl.watch.StackFrameDescriptorImpl;
+import com.intellij.ui.components.JBCheckBox;
 import com.intellij.ui.components.JBList;
+import com.intellij.ui.table.JBTable;
+import com.intellij.util.ui.table.JBListTable;
 import com.kkyeer.debugger.to.uml.helper.ImageType;
 import com.kkyeer.debugger.to.uml.helper.InvocationToImage;
 import com.kkyeer.debugger.to.uml.model.Invocation;
@@ -10,7 +13,12 @@ import com.kkyeer.debugger.to.uml.model.InvokeChain;
 import com.kkyeer.debugger.to.uml.model.InvokeType;
 
 import javax.swing.*;
+import javax.swing.event.ChangeEvent;
+import javax.swing.event.ChangeListener;
+import javax.swing.plaf.ListUI;
 import java.awt.*;
+import java.awt.event.ItemEvent;
+import java.awt.event.ItemListener;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -38,11 +46,25 @@ public class StackFrameControlPanel {
         JBList<JavaStackFrame> listUI = new JBList<>();
         JavaStackFrame[] frames = stackFrameList.toArray(new JavaStackFrame[stackFrameList.size()]);
         listUI.setListData(frames);
+        listUI.setFixedCellHeight(35);
         listUI.setCellRenderer(
                 (list, value, index, isSelected, cellHasFocus) -> {
-                    JLabel label = new JLabel();
-                    label.setText(value.getDescriptor().toString());
-                    return label;
+                    String text = value.getDescriptor().toString();
+                    JBCheckBox checkBox = new JBCheckBox();
+                    checkBox.setText(text);
+                    checkBox.setSelected(isSelected);
+                    checkBox.addItemListener(
+                            e -> System.out.println(e)
+                    );
+                    checkBox.addChangeListener(
+                            new ChangeListener() {
+                                @Override
+                                public void stateChanged(ChangeEvent e) {
+                                    System.out.println(e);
+                                }
+                            }
+                    );
+                    return checkBox;
                 }
         );
         listUI.setSelectionMode(ListSelectionModel.SINGLE_INTERVAL_SELECTION);
@@ -65,16 +87,16 @@ public class StackFrameControlPanel {
         List<Invocation> invocationList = new ArrayList<>();
         String prevClass = null;
         Invocation prevInvocation = null;
-        int max,min;
+        List<JavaStackFrame> selected = new ArrayList<>();
         if (selectedIndexes != null && selectedIndexes.length != 0) {
-            min = selectedIndexes[0];
-            max = selectedIndexes[selectedIndexes.length - 1];
+            for (int i = 0; i < selectedIndexes.length; i++) {
+                selected.add(items.get(selectedIndexes[i]));
+            }
         } else {
-            max = items.size() - 1;
-            min = 0;
+            selected = items;
         }
-        for (int i = max; i >= min; i--) {
-            JavaStackFrame frame = items.get(i);
+        for (int i = selected.size() - 1; i >= 0; i--) {
+            JavaStackFrame frame = selected.get(i);
             StackFrameDescriptorImpl descriptor = frame.getDescriptor();
             String currentClassName = getShortClassName(descriptor.getLocation().declaringType().name());