+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ## [1.0.0]
 ### Added
 - first version,This plugin allows you to generate sequence diagram uml from Java debugger stack and save it as svg file.
+
+## [1.0.1]
+### Support control for frames
+- now can choose some stack frames